pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-coherence"
version = "0.2.0"
description = "Token optimization layer for multi-agent LangGraph systems — 37–69% token savings via MESI cache coherence"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Arbiter contributors" }]
keywords = ["multi-agent", "llm", "cache-coherence", "mesi", "token-efficiency", "agents"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0"]

[project.scripts]
ccs-simulate = "ccs.cli.simulate:main"
ccs-compare = "ccs.cli.compare:main"
ccs-check-architecture = "ccs.hardening.architecture:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]
langgraph = ["langgraph>=0.2"]
crewai = ["crewai>=0.80"]
otel = ["opentelemetry-api>=1.0"]
langsmith = ["langsmith>=0.1"]
benchmark = ["langchain-core>=0.2"]
all = ["agent-coherence[langgraph,crewai,otel,langsmith,benchmark]"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools]
package-dir = { "" = "src" }
license-files = []

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
Homepage = "https://github.com/hipvlady/agent-coherence"
Repository = "https://github.com/hipvlady/agent-coherence"
Issues = "https://github.com/hipvlady/agent-coherence/issues"
